@@ -1,13 +1,14 @@
 package sa.edu.kau.fcit.cpit252;
 
-import org.junit.Test;
+
 import sa.edu.kau.fcit.cpit252.logging.Logger;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class Lab2Test {
 
