@@ -1,10 +1,9 @@
 package sa.edu.kau.fcit.cpit252;
 
-import org.junit.Test;
 import sa.edu.kau.fcit.cpit252.logging.Logger;
 
-import static org.junit.Assert.assertEquals;
-
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
 public class LoggerTest {
 
     @Test
